Add buttons for easier play
I added 3 buttons for choosing rock, paper or scissors and rewrote the
playRound function to take the button value as parameter.
The function then makes a random computer guess and compares it to the
button value.
Then a message is display showing the result and below is the score
which is updated after every round.
Once either the player or computer score reaches 5, an alert window
calls the winner of the game.